Move effects consts to constants
